requirements-pi.txt
ollama>=0.3.3
RPi.GPIO>=0.7.1
pumpfunpy>=0.2.3
playwright>=1.46.0

